Clean up button styles
