+++ frontend/src/theme/palette.ts
@@ -1,0 +1,51 @@
+import { PaletteOptions } from "@mui/material/styles";
+
+const palette: PaletteOptions = {
+  Learner: {
+    Light: "#F5FDFF",
+    Hover: "#CCF5FF",
+    Default: "#006877",
+    Pressed: "#00363F",
+  },
+  Administrator: {
+    Light: "#FFF8F6",
+    Hover: "#FFD9CC",
+    Default: "#8F4C34",
+    Pressed: "#5C1900",
+  },
+  Facilitator: {
+    Light: "#F9F5FF",
+    Hover: "#D6D6FF",
+    Default: "#555A92",
+    Pressed: "#1F257A",
+  },
+  Error: {
+    Light: "#FFF2F0",
+    Hover: "#FFD1CC",
+    Default: "#BA1A1A",
+    Pressed: "#690005",
+  },
+  Success: {
+    Light: "#FDFFF0",
+    Hover: "#FAFFCC",
+    Default: "#687021",
+    Pressed: "#444B04",
+  },
+  Warning: {
+    Light: "#FFFBEF",
+    Hover: "#FFF2CC",
+    Default: "#775900",
+    Pressed: "#3F2F00",
+  },
+  Neutral: {
+    100: "#FFFFFF",
+    200: "#F8FAFA",
+    300: "#E4E5E5",
+    400: "#CACCCC",
+    500: "#6F797B",
+    600: "#404B4D",
+    700: "#111111",
+  },
+};
+
+export default palette;
