+++ frontend/src/theme/typography.ts
@@ -1,0 +1,119 @@
+import { TypographyOptions } from "@mui/material/styles";
+
+const typography: TypographyOptions = {
+  displayLarge: {
+    fontSize: "46px",
+    fontWeight: 700,
+    lineHeight: "110%",
+    letterSpacing: "-0.2px",
+    textTransform: "none",
+  },
+  displayMedium: {
+    fontSize: "36px",
+    fontWeight: 700,
+    lineHeight: "110%",
+    textTransform: "none",
+  },
+  displaySmall: {
+    fontSize: "28px",
+    fontWeight: 700,
+    lineHeight: "110%",
+    textTransform: "none",
+  },
+  headlineLarge: {
+    fontSize: "28px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    textTransform: "none",
+  },
+  headlineMedium: {
+    fontSize: "26px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    textTransform: "none",
+  },
+  headlineSmall: {
+    fontSize: "22px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    textTransform: "none",
+  },
+  titleLarge: {
+    fontSize: "24px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    textTransform: "none",
+  },
+  titleMedium: {
+    fontSize: "18px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    letterSpacing: "0.12px",
+    textTransform: "none",
+  },
+  titleSmall: {
+    fontSize: "16px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    letterSpacing: "0.08px",
+    textTransform: "none",
+  },
+  bodyLarge: {
+    fontSize: "18px",
+    fontWeight: 400,
+    lineHeight: "140%",
+    letterSpacing: "0.4px",
+    textTransform: "none",
+  },
+  bodyMedium: {
+    fontSize: "16px",
+    fontWeight: 400,
+    lineHeight: "140%",
+    letterSpacing: "0.2px",
+    textTransform: "none",
+  },
+  bodySmall: {
+    fontSize: "14px",
+    fontWeight: 400,
+    lineHeight: "140%",
+    letterSpacing: "0.32px",
+    textTransform: "none",
+  },
+  labelLargeProminent: {
+    fontSize: "16px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    letterSpacing: "0.08px",
+    textTransform: "none",
+  },
+  labelLarge: {
+    fontSize: "14px",
+    fontWeight: 300,
+    lineHeight: "120%",
+    letterSpacing: "0.7px",
+    textTransform: "uppercase",
+  },
+  labelMediumProminent: {
+    fontSize: "14px",
+    fontWeight: 600,
+    lineHeight: "120%",
+    letterSpacing: "0.4px",
+    textTransform: "none",
+  },
+  labelMedium: {
+    fontSize: "14px",
+    fontWeight: 300,
+    lineHeight: "120%",
+    letterSpacing: "0.7px",
+    textTransform: "uppercase",
+  },
+  labelSmall: {
+    fontSize: "12.5px",
+    fontWeight: 300,
+    lineHeight: "120%",
+    letterSpacing: "0.625px",
+    textTransform: "uppercase",
+  },
+};
+
+export default typography;
